requirements.txt
python-dotenv
gradio_client==1.5.4
gradio==5.12.0
openai==1.37.0
# google-genai
google.generativeai
httpx==0.27.2
tiktoken
tinydb
tinymongo
proglog
yt-dlp>=2025.1.12
torch
torchaudio
### whisper timestamped 
whisper-timestamped
protobuf==3.20.3
pillow==10.4.0
moviepy==2.1.2
progress
questionary
edge-tts
